adding yaml output option, updating license to refer to "Kube Capacity Authors", adding contributors section
