+++ README.md
@@ -111,5 +111,12 @@ There are already some great projects out there that have similar goals.
 - [kube-resource-report](https://github.com/hjacobs/kube-resource-report): generates HTML/CSS report for resource requests and limits across multiple clusters.
 - [kubetop](https://github.com/LeastAuthority/kubetop): a CLI similar to top for Kubernetes, focused on resource utilization (not requests and limits).
 
+## Contributors
+
+Although this project was originally developed by [robscott](https://github.com/robscott), there have been some great contributions from others:
+
+- [endzyme](https://github.com/endzyme)
+- [justinbarrick](https://github.com/justinbarrick)
+
 ## License
 Apache License 2.0
