+++ pkg/capacity/table.go
@@ -1,4 +1,17 @@
-// Package capacity - text.go contains all the messy details for the text printer implementation
+// Copyright 2019 Kube Capacity Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package capacity
 
 import (
@@ -15,7 +28,7 @@ type tablePrinter struct {
 	w        *tabwriter.Writer
 }
 
-func (tp tablePrinter) Print() {
+func (tp *tablePrinter) Print() {
 	tp.w.Init(os.Stdout, 0, 8, 2, ' ', 0)
 	names := make([]string, len(tp.cm.nodeMetrics))
